Move FindAddrOpts and Symbol{Info,Type} into inspect module
This change moves the FindAddrOpts and Symbol{Info,Type} types into the
inspect module. They are specific to the inspect functionality, though
currently they are used in many parts of the crate because of the
current layering (basically everything goes through SymResolver).
