+++ src/symbolize/symbolizer.rs
@@ -6,6 +6,9 @@ use std::path::PathBuf;
 use crate::elf::ElfCache;
 use crate::elf::ElfResolver;
 use crate::gsym::GsymResolver;
+use crate::inspect::FindAddrOpts;
+use crate::inspect::SymbolInfo;
+use crate::inspect::SymbolType;
 use crate::kernel::KernelResolver;
 use crate::ksym::KSymCache;
 use crate::ksym::KALLSYMS;
@@ -17,11 +20,8 @@ use crate::normalize::UserAddrMeta;
 use crate::resolver::ResolverMap;
 use crate::util::uname_release;
 use crate::Addr;
-use crate::FindAddrOpts;
 use crate::Pid;
 use crate::SymResolver;
-use crate::SymbolInfo;
-use crate::SymbolType;
 
 use super::source::Elf;
 use super::source::Gsym;
